pyproject.toml
[project]
name = "glee"
version = "0.3.0"
description = "The Conductor for Your AI Orchestra"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # CLI
    "typer>=0.15.0",
    "rich>=14.0.0",
    # Orchestration
    "langgraph>=1.0.5",
    # Storage
    "lancedb>=0.26.1",
    "duckdb>=1.2.0",
    "fastembed>=0.7.4",
    # Types & Validation
    "pydantic>=2.12.5",
    "jsonschema>=4.26.0",
    # Logging
    "loguru>=0.7.0",
    # Config
    "pyyaml>=6.0.0",
    # MCP Server
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
glee = "glee.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["glee/**"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
]
